See issue #27: Re-organized files - Switched from execSync to spawnSync for large stdio (from help-all)
