+++ src/extension.ts
@@ -1,22 +1,12 @@
-import * as vscode from "vscode";
 import * as proc from "child_process";
-import * as path from "path";
+import * as vscode from "vscode";
 import { TargetsProvider, SourceRootsProvider } from "./treeviews";
 import { logger } from "./logging";
 import { CodeLensProvider } from "./codelens";
 import { TestsProvider } from "./test-explorer";
 import { getPantsExecutable } from "./configuration";
-import {
-  CancellationToken,
-  ConfigurationParams,
-  ConfigurationRequest,
-  DidChangeConfigurationNotification,
-  LanguageClient,
-  LanguageClientOptions,
-  ResponseError,
-  ServerOptions,
-  TransportKind,
-} from "vscode-languageclient/node";
+import { createLanguageClient, generateBuiltinsFile } from "./lsp/client";
+import { DidChangeConfigurationNotification, LanguageClient } from "vscode-languageclient/node";
 
 // TODO: Destructure vscode imports
 let lspClient: LanguageClient;
@@ -57,13 +47,17 @@ export async function activate(context: vscode.ExtensionContext) {
 
   const sourceRootsProvider = new SourceRootsProvider(rootPath);
   vscode.window.registerTreeDataProvider("source-roots", sourceRootsProvider);
-  vscode.commands.registerCommand("suspenders.source-roots.refresh", () =>
-    sourceRootsProvider.refresh()
+  context.subscriptions.push(
+    vscode.commands.registerCommand("suspenders.source-roots.refresh", () =>
+      sourceRootsProvider.refresh()
+    )
   );
 
   const targetsProvider = new TargetsProvider(rootPath);
   vscode.window.registerTreeDataProvider("targets", targetsProvider);
-  vscode.commands.registerCommand("suspenders.targets.refresh", () => targetsProvider.refresh());
+  context.subscriptions.push(
+    vscode.commands.registerCommand("suspenders.targets.refresh", () => targetsProvider.refresh())
+  );
 
   vscode.commands.executeCommand("setContext", "suspenders:isActivated", true);
 
@@ -75,9 +69,21 @@ export async function activate(context: vscode.ExtensionContext) {
     "All of the Tests"
   );
   const testsProvider = new TestsProvider(testController);
-  vscode.commands.registerCommand("suspenders.tests.refresh", () => testsProvider.refresh());
+  context.subscriptions.push(
+    vscode.commands.registerCommand("suspenders.tests.refresh", () => testsProvider.refresh())
+  );
 
-  // Hardcoding this in as a POC
+  context.subscriptions.push(
+    vscode.commands.registerCommand("suspenders.generateBuiltins", async () => {
+      try {
+        await generateBuiltinsFile(rootPath);
+      } catch (e) {
+        logger.error(`Error generating builtins file: ${e}`);
+      }
+    })
+  );
+
+  // Hardcoding this in as a POC - TODO Pull into LSP layer
   lspClient = createLanguageClient(context);
   await lspClient.start(); // Throwing away the promise return
   context.subscriptions.push(lspClient);
@@ -91,72 +97,6 @@ export function deactivate() {
   //   disposables.forEach((item) => item.dispose());
   // }
   // disposables = [];
-}
-
-// TODO: Grabbed this from https://github.com/microsoft/pyright/blob/496e50f65c8d3aecc43dcbc9b960d633cafe0c94/packages/vscode-pyright/src/extension.ts#L96
-// Spend some time cleaning it up for our purposes
-function createLanguageClient(context: vscode.ExtensionContext) {
-  let serverModule = context.asAbsolutePath(
-    path.join("node_modules", "pyright", "langserver.index.js")
-  );
-  const debugOptions = { execArgv: ["--nolazy", "--inspect=6600"] };
-
-  // TODO: Is stdio necessary? I ran into some errors using node ipc, but maybe those were false positives
-  const serverOptions: ServerOptions = {
-    run: { module: serverModule, transport: TransportKind.stdio, args: ["--stdio"] },
-    debug: {
-      module: serverModule,
-      transport: TransportKind.stdio,
-      args: ["--stdio"],
-      options: debugOptions,
-    },
-  };
-
-  const clientOptions: LanguageClientOptions = {
-    // TODO: this can also be a pattern on BUILD files, but I feel like language-based is more "correct"
-    documentSelector: [{ scheme: "file", language: "pantsbuild" }],
-    middleware: {
-      workspace: {
-        configuration: async (
-          params: ConfigurationParams,
-          token: CancellationToken,
-          next: ConfigurationRequest.HandlerSignature
-        ) => {
-          let result = next(params, token);
-
-          if (isThenable(result)) {
-            logger.debug("createLanguageClient: Result is Thenable");
-            result = await result;
-          }
-          logger.debug(
-            `createLanguageClient: Params: ${JSON.stringify(params)} - Result: ${JSON.stringify(result)}`
-          );
-
-          if (result instanceof ResponseError) {
-            logger.error(`createLanguageClient: ResponseError: ${JSON.stringify(result)}`);
-            return result;
-          }
-
-          for (const [i, item] of params.items.entries()) {
-            logger.debug(`Entries: ${i} ${JSON.stringify(item)}`);
-
-            // TODO: Hardcoding this as a proof-of-concept, but this should probably be part of Suspenders config
-            if (item.section === "python.analysis") {
-              (result[i] as any).stubPath = "./.pants.d";
-            }
-          }
-
-          return result;
-        },
-      },
-    },
-  };
-
-  return new LanguageClient("PantsBuild LSP Client", serverOptions, clientOptions);
-}
-
-function isThenable<T>(v: any): v is Thenable<T> {
-  return typeof v?.then === "function";
 }
 
 // TODO: Kibosh this - many better ways coming up than this using the new `peek`
